Начата вывод результатов скрапинга в браузер.

+++ models/realty_object_info_model.go
@@ -3,7 +3,7 @@ package models
 type ObjectInfo struct {
 	NotificationNumber string  // Номер извещения
 	Area               float64 // Площадь, кв.м
-	MonthlyRental      float64 // Стоимость аренды в месяц, рублей
+	MonthlyRental      int     // Стоимость аренды в месяц, рублей
 	RentalPeriod       int     // Срок аренды, месяцев TODO: на странице и 'лет', и 'мес.'
 	WebLink            string  // Ссылка для просмотра
 }
